+++ demo/src/main/java/cn/jzvd/demo/MainActivity.java
@@ -18,9 +18,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import cn.jzvd.Jzvd;
-import cn.jzvd.demo.fragment.BasicFragment;
-import cn.jzvd.demo.fragment.ComplexFragment;
 import cn.jzvd.demo.fragment.CustomFragment;
+import cn.jzvd.demo.fragment.ComplexFragment;
+import cn.jzvd.demo.fragment.BaseFragment;
 import cn.jzvd.demo.fragment.OtherFragment;
 
 import static cn.jzvd.Jzvd.backPress;
@@ -56,7 +56,7 @@ public class MainActivity extends AppCompatActivity {
         fragments = new ArrayList<>(4);
         items = new SparseIntArray(4);
 
-        BasicFragment basicsFragment = new BasicFragment();
+        BaseFragment basicsFragment = new BaseFragment();
         CustomFragment customFragment = new CustomFragment();
         ComplexFragment complexFragment = new ComplexFragment();
         OtherFragment otherFragment = new OtherFragment();
