--- demo/src/main/java/cn/jzvd/demo/api/BigChangeUiActivity.java
@@ -1,58 +1,0 @@
-package cn.jzvd.demo.api;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.MenuItem;
-import android.view.View;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
-import cn.jzvd.Jzvd;
-import cn.jzvd.demo.api.BigUIChangeAG.UiBigChangeAGActivity;
-import cn.jzvd.demo.R;
-
-/**
- * Created by Nathen on 16/7/31.
- */
-public class BigChangeUiActivity extends AppCompatActivity {
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        getSupportActionBar().setDisplayShowHomeEnabled(true);
-        getSupportActionBar().setDisplayShowTitleEnabled(true);
-        getSupportActionBar().setDisplayUseLogoEnabled(false);
-        getSupportActionBar().setTitle(getString(R.string.big_ui_change));
-        setContentView(R.layout.activity_ui_big_change);
-
-    }
-
-    @Override
-    public void onBackPressed() {
-        if (Jzvd.backPress()) {
-            return;
-        }
-        super.onBackPressed();
-    }
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-        Jzvd.releaseAllVideos();
-    }
-
-    public void clickUiAGBigChange(View view) {
-        startActivity(new Intent(BigChangeUiActivity.this, UiBigChangeAGActivity.class));
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case android.R.id.home:
-                finish();
-                break;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-}
