+++ demo/src/main/java/cn/jzvd/demo/fragment/BaseFragment.java
@@ -1,0 +1,134 @@
+package cn.jzvd.demo.fragment;
+
+import android.content.Intent;
+import android.hardware.Sensor;
+import android.hardware.SensorManager;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import com.bumptech.glide.Glide;
+
+import java.util.LinkedHashMap;
+
+import cn.jzvd.JZDataSource;
+import cn.jzvd.JZUtils;
+import cn.jzvd.Jzvd;
+import cn.jzvd.JzvdStd;
+import cn.jzvd.demo.ApplicationDemo;
+import cn.jzvd.demo.R;
+import cn.jzvd.demo.Urls;
+import cn.jzvd.demo.api.CustomMediaActivity;
+import cn.jzvd.demo.api.ExtendsNormalActivity;
+import cn.jzvd.demo.api.GetGifActivity;
+import cn.jzvd.demo.api.OrientationActivity;
+import cn.jzvd.demo.api.PreloadingActivity;
+import cn.jzvd.demo.api.RotationVideoSizeActivity;
+import cn.jzvd.demo.api.ScreenRotateActivity;
+
+import static android.content.Context.SENSOR_SERVICE;
+
+/**
+ * Created by pengan.li on 2020/5/8.
+ * 展示饺子的一些自定义用法，修改jzvdStd暴露的函数和变量，不需要继承jzvdStd的用法
+ */
+public class BaseFragment extends Fragment implements View.OnClickListener {
+
+    private JzvdStd mJzvdStd;
+    private Button mOrientation, mExtendsNormalActivity,
+            mRotationAndVideoSize, mCustomMediaPlayer, mPreLoading, mScreenRotate;
+    private Jzvd.JZAutoFullscreenListener mSensorEventListener;
+    private SensorManager mSensorManager;
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View view = View.inflate(getContext(), R.layout.fragment_api, null);
+        mJzvdStd = view.findViewById(R.id.jz_video);
+        mOrientation = view.findViewById(R.id.orientation);
+        mExtendsNormalActivity = view.findViewById(R.id.extends_normal_activity);
+        mRotationAndVideoSize = view.findViewById(R.id.rotation_and_videosize);
+        mCustomMediaPlayer = view.findViewById(R.id.custom_mediaplayer);
+        mPreLoading = view.findViewById(R.id.preloading);
+        mScreenRotate = view.findViewById(R.id.screen_rotate);
+
+        mOrientation.setOnClickListener(this);
+        mExtendsNormalActivity.setOnClickListener(this);
+        mRotationAndVideoSize.setOnClickListener(this);
+        mCustomMediaPlayer.setOnClickListener(this);
+        mPreLoading.setOnClickListener(this);
+        mScreenRotate.setOnClickListener(this);
+
+        mSensorManager = (SensorManager) getContext().getSystemService(SENSOR_SERVICE);
+        mSensorEventListener = new Jzvd.JZAutoFullscreenListener();
+
+        return view;
+    }
+
+
+    @Override
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        LinkedHashMap map = new LinkedHashMap();
+        String proxyUrl = ApplicationDemo.getProxy(getContext()).getProxyUrl(Urls.clarities[0]);
+        map.put("高清", proxyUrl);
+        map.put("标清", Urls.clarities[1]);
+        map.put("普清", Urls.clarities[2]);
+        JZDataSource jzDataSource = new JZDataSource(map, "饺子不信");
+        jzDataSource.looping = true;
+        jzDataSource.currentUrlIndex = 2;
+        jzDataSource.headerMap.put("key", "value");//header
+        mJzvdStd.setUp(jzDataSource
+                , JzvdStd.SCREEN_NORMAL);
+        Glide.with(this).load(Urls.videoPosterList[0]).into(mJzvdStd.posterImageView);
+    }
+
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        Sensor accelerometerSensor = mSensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER);
+        mSensorManager.registerListener(mSensorEventListener, accelerometerSensor, SensorManager.SENSOR_DELAY_NORMAL);
+        //home back
+        Jzvd.goOnPlayOnResume();
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        mSensorManager.unregisterListener(mSensorEventListener);
+        JZUtils.clearSavedProgress(getContext(), null);
+        //home back
+        Jzvd.goOnPlayOnPause();
+    }
+
+    @Override
+    public void onClick(View view) {
+        switch (view.getId()) {
+            case R.id.orientation:
+                startActivity(new Intent(getContext(), OrientationActivity.class));
+                break;
+            case R.id.extends_normal_activity:
+                startActivity(new Intent(getContext(), ExtendsNormalActivity.class));
+                break;
+            case R.id.rotation_and_videosize:
+                startActivity(new Intent(getContext(), RotationVideoSizeActivity.class));
+                break;
+            case R.id.custom_mediaplayer:
+                startActivity(new Intent(getContext(), CustomMediaActivity.class));
+                break;
+            case R.id.preloading:
+                startActivity(new Intent(getContext(), PreloadingActivity.class));
+                break;
+            case R.id.screen_rotate:
+                startActivity(new Intent(getContext(), ScreenRotateActivity.class));
+                break;
+
+        }
+    }
+}
