--- demo/src/main/java/cn/jzvd/demo/fragment/BasicFragment.java
@@ -1,42 +1,0 @@
-package cn.jzvd.demo.fragment;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import com.bumptech.glide.Glide;
-
-import cn.jzvd.demo.CustomJzvd.MyJzvdStd;
-import cn.jzvd.demo.R;
-
-/**
- * Created by pengan.li on 2020/5/8.
- * 展示饺子最基本的用法
- */
-public class BasicFragment extends Fragment {
-
-    private MyJzvdStd myJzvdStd;
-
-    @Nullable
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = View.inflate(getContext(), R.layout.fragment_basic, null);
-        myJzvdStd = view.findViewById(R.id.jz_video);
-        return view;
-    }
-
-
-    @Override
-    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-
-        myJzvdStd.setUp("http://jzvd.nathen.cn/342a5f7ef6124a4a8faf00e738b8bee4/cf6d9db0bd4d41f59d09ea0a81e918fd-5287d2089db37e62345123a1be272f8b.mp4"
-                , "饺子快长大");
-        Glide.with(this).load("http://jzvd-pic.nathen.cn/jzvd-pic/1bb2ebbe-140d-4e2e-abd2-9e7e564f71ac.png").into(myJzvdStd.posterImageView);
-
-    }
-}
