+++ demo/src/main/java/cn/jzvd/demo/fragment/CustomFragment.java
@@ -1,8 +1,6 @@
 package cn.jzvd.demo.fragment;
 
 import android.content.Intent;
-import android.hardware.Sensor;
-import android.hardware.SensorManager;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -14,67 +12,61 @@ import androidx.fragment.app.Fragment;
 
 import com.bumptech.glide.Glide;
 
-import java.util.LinkedHashMap;
-
-import cn.jzvd.JZDataSource;
-import cn.jzvd.JZUtils;
 import cn.jzvd.Jzvd;
 import cn.jzvd.JzvdStd;
-import cn.jzvd.demo.ApplicationDemo;
+import cn.jzvd.demo.CustomJzvd.JzvdStdAutoCompleteAfterFullscreen;
+import cn.jzvd.demo.CustomJzvd.JzvdStdLockScreen;
+import cn.jzvd.demo.CustomJzvd.JzvdStdMp3;
+import cn.jzvd.demo.CustomJzvd.JzvdStdShowShareButtonAfterFullscreen;
+import cn.jzvd.demo.CustomJzvd.JzvdStdShowTextureViewAfterAutoComplete;
+import cn.jzvd.demo.CustomJzvd.JzvdStdShowTitleAfterFullscreen;
+import cn.jzvd.demo.CustomJzvd.JzvdStdSpeed;
+import cn.jzvd.demo.CustomJzvd.JzvdStdVolume;
+import cn.jzvd.demo.CustomJzvd.JzvdStdVolumeAfterFullscreen;
 import cn.jzvd.demo.R;
 import cn.jzvd.demo.Urls;
-import cn.jzvd.demo.api.BigChangeUiActivity;
-import cn.jzvd.demo.api.CustomMediaActivity;
-import cn.jzvd.demo.api.ExtendsNormalActivity;
-import cn.jzvd.demo.api.GetGifActivity;
-import cn.jzvd.demo.api.OrientationActivity;
-import cn.jzvd.demo.api.PreloadingActivity;
-import cn.jzvd.demo.api.RotationVideoSizeActivity;
-import cn.jzvd.demo.api.ScreenRotateActivity;
-import cn.jzvd.demo.api.SmallChangeUiActivity;
-
-import static android.content.Context.SENSOR_SERVICE;
+import cn.jzvd.demo.api.BigUIChangeAG.UiBigChangeAGActivity;
 
 /**
  * Created by pengan.li on 2020/5/8.
- * 展示饺子的一些自定义用法
+ * 展示饺子最基本的用法
  */
 public class CustomFragment extends Fragment implements View.OnClickListener {
 
-    private JzvdStd mJzvdStd;
-    private Button mSmallChange, mBigChange, mOrientation, mExtendsNormalActivity,
-            mRotationAndVideoSize, mCustomMediaPlayer, mPreLoading, mScreenRotate, mGetGif;
-    private Jzvd.JZAutoFullscreenListener mSensorEventListener;
-    private SensorManager mSensorManager;
+    JzvdStdShowShareButtonAfterFullscreen jzvdStdWithShareButton;
+    JzvdStdShowTitleAfterFullscreen jzvdStdShowTitleAfterFullscreen;
+    JzvdStdShowTextureViewAfterAutoComplete jzvdStdShowTextureViewAfterAutoComplete;
+    JzvdStdAutoCompleteAfterFullscreen jzvdStdAutoCompleteAfterFullscreen;
+    JzvdStdVolumeAfterFullscreen jzvdStdVolumeAfterFullscreen;
+    JzvdStdMp3 jzvdStdMp3;
+    JzvdStdSpeed jzvdStdSpeed;
+    JzvdStdLockScreen lockScreen;
+    JzvdStdVolume jzvdStdVolume;
+
+    JzvdStd jzvdStd_1_1, jzvdStd_16_9;
+    JzvdStd jzNoTitle;
+    Button agVideo;
 
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = View.inflate(getContext(), R.layout.fragment_api, null);
-        mJzvdStd = view.findViewById(R.id.jz_video);
-        mSmallChange = view.findViewById(R.id.small_change);
-        mBigChange = view.findViewById(R.id.big_change);
-        mOrientation = view.findViewById(R.id.orientation);
-        mExtendsNormalActivity = view.findViewById(R.id.extends_normal_activity);
-        mRotationAndVideoSize = view.findViewById(R.id.rotation_and_videosize);
-        mCustomMediaPlayer = view.findViewById(R.id.custom_mediaplayer);
-        mPreLoading = view.findViewById(R.id.preloading);
-        mScreenRotate = view.findViewById(R.id.screen_rotate);
-        mGetGif = view.findViewById(R.id.get_gif);
+        View view = View.inflate(getContext(), R.layout.fragment_custom, null);
 
-        mSmallChange.setOnClickListener(this);
-        mBigChange.setOnClickListener(this);
-        mOrientation.setOnClickListener(this);
-        mExtendsNormalActivity.setOnClickListener(this);
-        mRotationAndVideoSize.setOnClickListener(this);
-        mCustomMediaPlayer.setOnClickListener(this);
-        mPreLoading.setOnClickListener(this);
-        mScreenRotate.setOnClickListener(this);
-        mGetGif.setOnClickListener(this);
+        agVideo = view.findViewById(R.id.agvideo);
+        jzNoTitle = view.findViewById(R.id.jz_notitle);
+        lockScreen = view.findViewById(R.id.lock_screen);
+        jzvdStdWithShareButton = view.findViewById(R.id.custom_videoplayer_standard_with_share_button);
+        jzvdStdShowTitleAfterFullscreen = view.findViewById(R.id.custom_videoplayer_standard_show_title_after_fullscreen);
+        jzvdStdShowTextureViewAfterAutoComplete = view.findViewById(R.id.custom_videoplayer_standard_show_textureview_aoto_complete);
+        jzvdStdAutoCompleteAfterFullscreen = view.findViewById(R.id.custom_videoplayer_standard_aoto_complete);
+        jzvdStd_1_1 = view.findViewById(R.id.jz_videoplayer_1_1);
+        jzvdStd_16_9 = view.findViewById(R.id.jz_videoplayer_16_9);
+        jzvdStdVolumeAfterFullscreen = view.findViewById(R.id.jz_videoplayer_volume);
+        jzvdStdMp3 = view.findViewById(R.id.jz_videoplayer_mp3);
+        jzvdStdSpeed = view.findViewById(R.id.jz_videoplayer_speed);
+        jzvdStdVolume = view.findViewById(R.id.custom_videoplayer_standard_with_volume_button);
 
-        mSensorManager = (SensorManager) getContext().getSystemService(SENSOR_SERVICE);
-        mSensorEventListener = new Jzvd.JZAutoFullscreenListener();
-
+        agVideo.setOnClickListener(this);
         return view;
     }
 
@@ -82,69 +74,82 @@ public class CustomFragment extends Fragment implements View.OnClickListener {
     @Override
     public void onActivityCreated(@Nullable Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
-        LinkedHashMap map = new LinkedHashMap();
-        String proxyUrl = ApplicationDemo.getProxy(getContext()).getProxyUrl(Urls.clarities[0]);
-        map.put("高清", proxyUrl);
-        map.put("标清", Urls.clarities[1]);
-        map.put("普清", Urls.clarities[2]);
-        JZDataSource jzDataSource = new JZDataSource(map, "饺子不信");
-        jzDataSource.looping = true;
-        jzDataSource.currentUrlIndex = 2;
-        jzDataSource.headerMap.put("key", "value");//header
-        mJzvdStd.setUp(jzDataSource
-                , JzvdStd.SCREEN_NORMAL);
-        Glide.with(this).load(Urls.videoPosterList[0]).into(mJzvdStd.posterImageView);
+
+
+        jzNoTitle.setUp("http://jzvd.nathen.cn/342a5f7ef6124a4a8faf00e738b8bee4/cf6d9db0bd4d41f59d09ea0a81e918fd-5287d2089db37e62345123a1be272f8b.mp4"
+                , "饺子叫啥也显示不出来啊");
+        Glide.with(this).load("http://jzvd-pic.nathen.cn/jzvd-pic/1bb2ebbe-140d-4e2e-abd2-9e7e564f71ac.png").into(jzNoTitle.posterImageView);
+
+        lockScreen.setUp("http://jzvd.nathen.cn/342a5f7ef6124a4a8faf00e738b8bee4/cf6d9db0bd4d41f59d09ea0a81e918fd-5287d2089db37e62345123a1be272f8b.mp4"
+                , "饺子定身");
+        Glide.with(this).load("http://jzvd-pic.nathen.cn/jzvd-pic/1bb2ebbe-140d-4e2e-abd2-9e7e564f71ac.png").into(lockScreen.posterImageView);
+
+
+        jzvdStdWithShareButton.setUp(Urls.videoUrlList[3], "饺子想呼吸", Jzvd.SCREEN_NORMAL);
+        Glide.with(this)
+                .load(Urls.videoPosterList[3])
+                .into(jzvdStdWithShareButton.posterImageView);
+
+
+        jzvdStdShowTitleAfterFullscreen.setUp(Urls.videoUrlList[4], "饺子想摇头", Jzvd.SCREEN_NORMAL);
+        Glide.with(this)
+                .load(Urls.videoPosterList[4])
+                .into(jzvdStdShowTitleAfterFullscreen.posterImageView);
+
+        jzvdStdShowTextureViewAfterAutoComplete.setUp(Urls.videoUrlList[5], "饺子想旅行", Jzvd.SCREEN_NORMAL);
+        Glide.with(this)
+                .load(Urls.videoPosterList[5])
+                .into(jzvdStdShowTextureViewAfterAutoComplete.posterImageView);
+
+        jzvdStdAutoCompleteAfterFullscreen.setUp(Urls.videoUrls[0][1], "饺子没来", Jzvd.SCREEN_NORMAL);
+        Glide.with(this)
+                .load(Urls.videoPosters[0][1])
+                .into(jzvdStdAutoCompleteAfterFullscreen.posterImageView);
+
+        jzvdStd_1_1.setUp(Urls.videoUrls[0][1], "饺子有事吗", Jzvd.SCREEN_NORMAL);
+        Glide.with(this)
+                .load(Urls.videoPosters[0][1])
+                .into(jzvdStd_1_1.posterImageView);
+        jzvdStd_1_1.widthRatio = 1;
+        jzvdStd_1_1.heightRatio = 1;
+
+        jzvdStd_16_9.setUp(Urls.videoUrls[0][1], "饺子来不了", Jzvd.SCREEN_NORMAL);
+        Glide.with(this)
+                .load(Urls.videoPosters[0][1])
+                .into(jzvdStd_16_9.posterImageView);
+        jzvdStd_16_9.widthRatio = 16;
+        jzvdStd_16_9.heightRatio = 9;
+
+        jzvdStdVolumeAfterFullscreen.setUp(Urls.videoUrls[0][1], "饺子摇摆", Jzvd.SCREEN_NORMAL);
+        Glide.with(this)
+                .load(Urls.videoPosters[0][1])
+                .into(jzvdStdVolumeAfterFullscreen.posterImageView);
+
+        jzvdStdMp3.setUp(Urls.videoUrls[0][1],
+                "饺子你听", Jzvd.SCREEN_NORMAL);
+        Glide.with(this)
+                .load(Urls.videoPosters[0][1])
+                .into(jzvdStdMp3.posterImageView);
+
+
+        jzvdStdSpeed.setUp(Urls.videoUrls[0][1],
+                "饺子快点", Jzvd.SCREEN_NORMAL);
+        Glide.with(this)
+                .load(Urls.videoPosters[0][1])
+                .into(jzvdStdSpeed.posterImageView);
+
+
+        jzvdStdVolume.setUp(Urls.videoUrls[0][1],
+                "饺子吃莽莽", Jzvd.SCREEN_NORMAL);
+        Glide.with(this)
+                .load(Urls.videoPosters[0][1])
+                .into(jzvdStdVolume.posterImageView);
     }
 
-
     @Override
-    public void onResume() {
-        super.onResume();
-        Sensor accelerometerSensor = mSensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER);
-        mSensorManager.registerListener(mSensorEventListener, accelerometerSensor, SensorManager.SENSOR_DELAY_NORMAL);
-        //home back
-        Jzvd.goOnPlayOnResume();
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-        mSensorManager.unregisterListener(mSensorEventListener);
-        JZUtils.clearSavedProgress(getContext(), null);
-        //home back
-        Jzvd.goOnPlayOnPause();
-    }
-
-    @Override
-    public void onClick(View view) {
-        switch (view.getId()) {
-            case R.id.small_change:
-                startActivity(new Intent(getContext(), SmallChangeUiActivity.class));
-                break;
-            case R.id.big_change:
-                startActivity(new Intent(getContext(), BigChangeUiActivity.class));
-                break;
-            case R.id.orientation:
-                startActivity(new Intent(getContext(), OrientationActivity.class));
-                break;
-            case R.id.extends_normal_activity:
-                startActivity(new Intent(getContext(), ExtendsNormalActivity.class));
-                break;
-            case R.id.rotation_and_videosize:
-                startActivity(new Intent(getContext(), RotationVideoSizeActivity.class));
-                break;
-            case R.id.custom_mediaplayer:
-                startActivity(new Intent(getContext(), CustomMediaActivity.class));
-                break;
-            case R.id.preloading:
-                startActivity(new Intent(getContext(), PreloadingActivity.class));
-                break;
-            case R.id.screen_rotate:
-                startActivity(new Intent(getContext(), ScreenRotateActivity.class));
-                break;
-            case R.id.get_gif:
-                startActivity(new Intent(getContext(), GetGifActivity.class));
-                break;
+    public void onClick(View v) {
+        if (v.getId() == R.id.agvideo) {
+            startActivity(new Intent(getContext(), UiBigChangeAGActivity.class));
         }
     }
 }
