--- demo/src/main/res/layout/fragment_basic.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <cn.jzvd.demo.CustomJzvd.MyJzvdStd
-        android:id="@+id/jz_video"
-        android:layout_width="match_parent"
-        android:layout_height="200dp" />
-
-</LinearLayout>
