+++ demo/src/main/res/layout/fragment_complex.xml
@@ -21,12 +21,10 @@
         android:text="@string/tiny_window" />
 
     <Button
-        android:id="@+id/custom_ag_video"
+        android:id="@+id/get_gif"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="center_horizontal"
-        android:layout_marginTop="8dp"
         android:layout_marginBottom="8dp"
-        android:text="@string/custom_ag" />
-
+        android:text="@string/get_gif" />
 </LinearLayout>
