+++ demo/src/main/res/layout/fragment_custom.xml
@@ -13,6 +13,15 @@
             android:layout_height="wrap_content"
             android:orientation="vertical">
 
+            <Button
+                android:id="@+id/agvideo"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center_horizontal"
+                android:layout_marginTop="8dp"
+                android:layout_marginBottom="8dp"
+                android:text="@string/agplayer" />
+
             <TextView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
