+++ demo/src/main/res/values-zh/strings.xml
@@ -5,15 +5,14 @@
     <string name="custom">自定义</string>
     <string name="complex">复杂</string>
     <string name="other">其他</string>
-    <string name="small_change">小改饺子播放器(自定义)</string>
     <string name="big_ui_change">大改饺子播放器(自定义)</string>
     <string name="orientation">视频播放方向控制</string>
     <string name="extend_activity">继承Activity非AppCompatActivity</string>
     <string name="video_cover_and_orientation">调整视频填充方式、播放方向</string>
     <string name="custom_mediaplayer">自定义播放内核</string>
-    <string name="preloading">视频缓存</string>
+    <string name="preloading">视频预加载</string>
     <string name="screen_rotate">视频根据传感器自动旋转</string>
-    <string name="get_gif">获取视频截图</string>
+    <string name="get_gif">Gif截图</string>
     <string name="extend_jzvdstd_delete_view">继承JzvdStd删除父布局的一些控件</string>
     <string name="jzvdstd_delete_view_add_lock">继承JzvdStd删除父布局的一些控件并添加锁定按钮</string>
     <string name="add_share_btn_after_fullscreen">JzvdStd上添加分享按钮全屏播放时显示</string>
